Finished stockDownloader and added more functionality to Stock class
Modified 30/11/2016:
PortfolioTracker
* On initialization now creates a table in the database for holding
historical
price information for the stocks.
* Implemented getValue() and getCost() methods to calculate the total
value and cost of the portfolio on specific date.
Modified 30/11/2016:
* Class now downloads all past stock price data on initialization. Or
alternatively, if the stock has been added to the database previously,
it
updates the database with the most recent data.
* Implemented getPrice() and getValue() methods to get the price of the
stock
on a particular date, and the total value of the stock on a particular
date.
* Implemented getPriceRange() method which gets the price of the stock
over a
range of dates.
* Implemented getValueRange() method which gets a data frame containing
Date,
Total_Owned and Total_Value for a range of dates.